Dev_mate/requirements.txt
# ──────────────────────────────────────────────────────────
# DEVMATE – Python Dependencies
# Install with:  pip install -r requirements.txt
# ──────────────────────────────────────────────────────────

# ── LLM ──────────────────────────────────────────────────
ollama>=0.2.0                   # Ollama Python client (local LLM)

# ── Speech-to-Text ───────────────────────────────────────
openai-whisper>=20231117        # Offline STT (base model = 145 MB)
SpeechRecognition>=3.10.0       # Microphone capture helper

# ── Text-to-Speech ───────────────────────────────────────
TTS>=0.22.0                     # Coqui TTS (VITS model)

# ── Audio I/O ────────────────────────────────────────────
pyaudio>=0.2.13                 # Microphone recording
sounddevice>=0.4.6              # Audio playback (cross-platform)
soundfile>=0.12.1               # Read/write WAV files

# ── Time & Scheduling ────────────────────────────────────
python-dateutil>=2.9.0          # Human time string parsing

# ── Utilities ────────────────────────────────────────────
pillow>=10.0.0                  # Image support for Tkinter assets

# ── Development / Testing ────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.12.0

# ── Google Workspace ─────────────────────────────────────
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# ── Face Recognition & Automation ────────────────────────
opencv-python>=4.8.0
face-recognition>=1.3.0
pyautogui>=0.9.54
selenium>=4.15.0
webdriver-manager>=4.0.1
